User Form successfully being saved to db
Blog/Portfolio create feature is successfully storing the topic and the text to db.
