--- Project/Website/forms.py
@@ -1,9 +1,0 @@
-from typing import Text
-from flask_wtf import FlaskForm
-from wtforms import StringField,SubmitField,TextAreaField
-from wtforms.validators import DataRequired
-
-class BlogPostForm(FlaskForm):
-    title = StringField('Title', validators=[DataRequired()])
-    text = TextAreaField('Text', validators=[DataRequired()])
-    submit = SubmitField("Post")
