+++ Project/Website/templates/create_post.html
@@ -4,9 +4,11 @@
     <form class="form-group" method="POST">
         {{form.hidden_tag()}}
 
+    
         <div class="form-group">
             {{form.title.label(class="form-control-label")}}
             {{form.title(class="form-control form-control-lg")}}
+            
 
         </div>
 
@@ -17,7 +19,7 @@
         </div>
 
         <div class="form-group">
-            [{form.submit(class="btn btn-outline-info")}]
+            {{form.submit(class="btn btn-outline-info")}}
         </div>
 
     </form>
